wifi: make with -j4 option leads makes softap driver not embedded in filesystem
BZ: 4095

We've discovered that when building the project with the "-j4" option, some
synchronization issues might occur.
In our case, the multiple call to kenel-build.sh script make the install
process unsuccessful for the softap driver (not exported into the filesystem).

It is resolved in this script by using dynamic temporary file names in order to
make build threads deal with different file names.

Change-Id: I6cadae2f0e44b8a7e7f3e20c55ab1bbafb3ec7e8
Signed-off-by: Jeremie GARCIA <jeremiex.garcia@intel.com>