wl12xx-compat: Clean the temoprary output folder.
BZ: 11079

The kernel modules related to MLCP driver are generated with a bad
magic version number. This occurs when a patch is applied on the kernel
while the wl12xx-compat driver remain unchanged.

When the kernel magic number of the kernel modules are not the same as
the uts_release, then the wifi driver could not be loaded successfully.
The easiest way to resolve this issue is to remove the temporary
destination directory.

Change-Id: If0ff7ffc1ebb04bf3000da59e90438c1d08ad0be
Signed-off-by: Aymen Zayet <aymenx.zayet@intel.com>
Reviewed-on: http://android.intel.com:8080/20524
Reviewed-by: Champciaux, NicolasX <nicolasx.champciaux@intel.com>
Tested-by: Champciaux, NicolasX <nicolasx.champciaux@intel.com>
Reviewed-by: Monnier, OlivierX <olivierx.monnier@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>