MUST_REVERT kernel.mk: Fool Kbuild into using minigzip instead of gzip
BZ: 26498

We need to compress the bzImage with minigzip and not gzip.
This is because the deflate algorithm in gzip is very old compared
to zlib, and the recovery imgdiff codes create very large diffs
if the source file isn't compressed with the same version of deflate
that it has access to.  A wrapper hack script is added to the PATH
when invoking Kbuild to fool it into using minigzip, which is built
in external/zlib and staged with the host tools in the PATH.

This is a gross hack which should be reverted once we have a
cleaner solution.

Change-Id: If3f11f70d2530b118554083393cb06592538df38
Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>
Reviewed-on: http://android.intel.com:8080/38385
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>