publish_build: rename of flash xml
BZ:61284

name flash-no-modem confuse people. On variant that has no modem there should only be one flash xml file. It is more logical to call it flash.xml. For the one that has a modem the no modem xml file should be there only to reflash the board after failing flashing it with the appropriate xml file.
Signed-off-by: Marie-Cecile Afantenos <marie.cecilex.afantenos@intel.com>

Change-Id: I0bd2424496609a2820cc1cef5e712777050f5b39
Reviewed-on: http://android.intel.com:8080/69525
Reviewed-by: Jauffres, MichelX <michelx.jauffres@intel.com>
Reviewed-by: Romieu, Benoit <benoit.romieu@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Tested-by: Afantenos, Marie CecileX <marie.cecilex.afantenos@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>