Changed where and how size_hint is printed into recovery.fstab
BZ: 89060

(patch re-submitted as a first re-integration step to avoid FOTA issue)
That change was made to generate a recovery.fstab that is standard,
and not an Intel customized version to support MBR partitioning.
With that change, the size_hint is specified as a comment, which
makes it still usable for Intel, and ignored by default Android code.

Change-Id: I2e8bd52c8d5d4ced84ca724c067bf5a4403c8159
Signed-off-by: Florent Auger <florent.auger@intel.com>