Add ULPMC flashing by default in fastboot flashfile
BZ: 117275

Flash ulpmc will be trigged if ULPMC_BIANRY environnement variable
points to a existing file.
The fastboot flash ulpmc command is not mandatory, this means that if
the command fails, the flash procedure will still continue.

PhoneFlashTool 4.3.1.0 and later is required to support the mandatory
flag. For older PhoneFlashTool, the flag will be ignored and all
fastboot command failure will stop the flashing procedure as usual.

Change-Id: If4dc3b82a6f059b079c95ca852902ade84b8166d
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: http://android.intel.com:8080/117985
Reviewed-by: Despalles, Frederic <frederic.despalles@intel.com>
Reviewed-by: Auger, Florent <florent.auger@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Tested-by: Molero Rodenas, BeatrizX <beatrizx.molero.rodenas@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>