Do not generate the external release by default
BZ: 159587

If EXTERNAL_BINARIES is not set in the environment, the
default value is set to the opposite of FLASHFILE_BOOTONLY
which is set to FALSE by default, therefore EXTERNAL_BINARIES
is set to TRUE by default.

It should be set to FALSE instead, there is no point in building
the external binaries unless explicitely requested (and besides,
it breaks the external build itself unless EXTERNAL_BINARIES is
explicitely set to FALSE as Buildbot does).

Change-Id: I66bbd5fb5ade8ffc29bdce01f10c3d2c6d9f89c9
Signed-off-by: Olivier Fourdan <olivier.fourdan@intel.com>