[REVERTME] Wipe reserved partition in UEFI blankphone
BZ: 170234

A partitioning layout change has moved ESP position without wiping its
content.
The result is that UEFI BIOS found the old ESP FS and use it instead of
the new one.

This patch is used to clear all the boards that have the residual ESP
and will have to be reverted once it's done.

Change-Id: Ia0c2403bd1814620f4e115704a9e21dc2503066d
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@gmail.com>