flashfile: skip missing targets
BZ: 193849

Some targets may be in the flashfile definition but not available at
build time. This is the case for the capsule which is available in
internal builds but not in external builds.

If a target from the flashfile description is not available, skip it and
do not add a command for it.

Change-Id: Ie7b5d6defd50033fd972608dfc52bf3e646dbbd9
Signed-off-by: Laurent FERT <laurent.fert@intel.com>