OTA cleanup : mkbootimg must be called from AOSP
When called from AOSP, some parameters are missing, like "type"
(type of OS to be generated). Try to guess from the image name
(possible on standard image generation), or decide it if impossible
(on OTA generation for example) - will be fixed when flashed.

Add "adb wait-for-device" after "adb root" to be independant from
PFT handling.

Change-Id: If06747057648ab66bd391f3e66da4a58b4217037
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1988
Signed-off-by: Guillaume Betous <guillaumex.betous@intel.com>
Reviewed-on: https://android.intel.com:443/265310